+++ ui/src/pages/chat/chat-send.test.ts
@@ -14,6 +14,7 @@ import {
   buildSlashCommandsFromEntries,
   replaceSlashCommands,
 } from "../../lib/chat/commands.ts";
+import { extractText } from "../../lib/chat/message-extract.ts";
 import { createResolvedModelPatch } from "../../test-helpers/chat-model.ts";
 import {
   createTestGatewayClient,
@@ -40,6 +41,7 @@ import {
 } from "./chat-session.ts";
 import { patchChatSessionSettings } from "./chat-settings-patches.ts";
 import type { ChatPageHost } from "./chat-state-host.ts";
+import { buildChatItems } from "./chat-thread-build.ts";
 import {
   admitStoredChatComposerQueueItem,
   listStoredChatOutboxes,
@@ -9217,9 +9219,14 @@ describe("handleSendChat", () => {
 
   it("removes queued steer indicators when chat.send returns terminal ok", async () => {
     const original = { id: "queued-1", text: "tighten the plan", createdAt: 1 };
+    let steerRequestRunId: string | undefined;
     const host = makeChatHost({
       requestHandlers: {
-        "chat.send": { status: "ok", runId: "steer-ok" },
+        "chat.send": (params: unknown) => {
+          const payload = requireRecord(params, "terminal steer payload");
+          steerRequestRunId = String(payload.idempotencyKey);
+          return { status: "ok", runId: "steer-ok" };
+        },
       },
       chatRunId: "run-1",
       chatDisplayedLeafEntryId: "leaf-active",
@@ -9232,7 +9239,39 @@ describe("handleSendChat", () => {
     await steerQueuedChatMessage(host, "queued-1");
 
     expect(host.chatRunId).toBe("run-1");
-    expect(host.chatStream).toBe("Working...");
+    expect(host.chatStream).toBeNull();
+    expect(host.chatStreamSegments).toEqual([
+      {
+        text: "Working...",
+        ts: expect.any(Number),
+        runId: "run-1",
+        boundaryRunId: steerRequestRunId,
+      },
+    ]);
+    expect(host.chatMessages).toEqual([
+      expect.objectContaining({
+        role: "user",
+        __openclaw: { idempotencyKey: `${steerRequestRunId}:user` },
+      }),
+    ]);
+    const renderedText = buildChatItems({
+      paneId: "terminal-steer",
+      sessionKey: host.sessionKey,
+      runId: host.chatRunId,
+      messages: host.chatMessages,
+      toolMessages: host.chatToolMessages,
+      streamSegments: host.chatStreamSegments,
+      stream: host.chatStream,
+      streamStartedAt: host.chatStreamStartedAt,
+      showToolCalls: true,
+    }).flatMap((item) =>
+      item.kind === "stream"
+        ? [item.text]
+        : item.kind === "group"
+          ? item.messages.map(({ message }) => extractText(message))
+          : [],
+    );
+    expect(renderedText).toEqual(["Working...", "tighten the plan"]);
     expect(host.chatQueue).toStrictEqual([]);
     expect(host.applySettings).toHaveBeenCalledWith(
       expect.objectContaining({ lastActiveSessionKey: "agent:main:main" }),
