+++ ui/src/pages/chat/split-layout.ts
@@ -1,5 +1,9 @@
 import type { UiCommand } from "@openclaw/gateway-protocol";
-import { expectDefined, isRecord } from "@openclaw/normalization-core";
+import { expectDefined } from "@openclaw/normalization-core";
+import {
+  normalizeSplitLayoutWeights,
+  splitLayoutNumericSuffix,
+} from "./split-layout-persistence.ts";
 
 export type ChatSplitPane = { id: string; sessionKey: string };
 type ChatSplitColumn = { id: string; panes: ChatSplitPane[]; paneWeights: number[] };
@@ -46,23 +50,9 @@ function cloneLayout(layout: ChatSplitLayout): ChatSplitLayout {
   };
 }
 
-function equalWeights(length: number): number[] {
-  return Array.from({ length }, () => 1 / length);
-}
-
-function normalizedWeights(weights: number[]): number[] {
-  const total = weights.reduce((sum, weight) => sum + weight, 0);
-  return weights.map((weight) => weight / total);
-}
-
-function numericSuffix(id: string, prefix: string): number {
-  const match = new RegExp(`^${prefix}(\\d+)$`, "u").exec(id);
-  return match ? Number(match[1]) : 0;
-}
-
 function nextColumnId(layout: ChatSplitLayout): string {
   const max = layout.columns.reduce(
-    (current, column) => Math.max(current, numericSuffix(column.id, "c")),
+    (current, column) => Math.max(current, splitLayoutNumericSuffix(column.id, "c")),
     0,
   );
   return `c${max + 1}`;
@@ -70,7 +60,7 @@ function nextColumnId(layout: ChatSplitLayout): string {
 
 function nextPaneId(layout: ChatSplitLayout): string {
   const max = panesOf(layout).reduce(
-    (current, pane) => Math.max(current, numericSuffix(pane.id, "p")),
+    (current, pane) => Math.max(current, splitLayoutNumericSuffix(pane.id, "p")),
     0,
   );
   return `p${max + 1}`;
@@ -181,12 +171,12 @@ export function closePane(layout: ChatSplitLayout, paneId: string): ChatSplitLay
     next.columns.splice(location.columnIndex, 1);
     next.columnWeights.splice(location.columnIndex, 1);
   } else {
-    column.paneWeights = normalizedWeights(column.paneWeights);
+    column.paneWeights = normalizeSplitLayoutWeights(column.paneWeights);
   }
   if (panesOf(next).length <= 1) {
     return undefined;
   }
-  next.columnWeights = normalizedWeights(next.columnWeights);
+  next.columnWeights = normalizeSplitLayoutWeights(next.columnWeights);
   next.activePaneId = nextActivePaneId;
   return next;
 }
@@ -281,111 +271,4 @@ export function resizePanes(
     column.paneWeights = resizePair(column.paneWeights, boundaryIndex, pairRatio);
   }
   return next;
-}
-
-function readWeights(value: unknown, length: number): number[] {
-  if (
-    !Array.isArray(value) ||
-    value.length !== length ||
-    value.some((weight) => typeof weight !== "number" || !Number.isFinite(weight) || weight <= 0)
-  ) {
-    return equalWeights(length);
-  }
-  return normalizedWeights(value);
-}
-
-function uniqueId(value: unknown, used: Set<string>, next: () => string): string {
-  const candidate = typeof value === "string" ? value.trim() : "";
-  if (candidate && !used.has(candidate)) {
-    used.add(candidate);
-    return candidate;
-  }
-  let generated = next();
-  while (used.has(generated)) {
-    generated = next();
-  }
-  used.add(generated);
-  return generated;
-}
-
-export function normalizeChatSplitLayout(value: unknown): ChatSplitLayout | undefined {
-  if (!isRecord(value) || !Array.isArray(value.columns)) {
-    return undefined;
-  }
-  const rawColumns = value.columns.filter(isRecord);
-  let paneSequence = rawColumns.reduce((max, rawColumn) => {
-    if (!Array.isArray(rawColumn.panes)) {
-      return max;
-    }
-    return rawColumn.panes.reduce((paneMax, rawPane) => {
-      if (!isRecord(rawPane) || typeof rawPane.id !== "string") {
-        return paneMax;
-      }
-      return Math.max(paneMax, numericSuffix(rawPane.id.trim(), "p"));
-    }, max);
-  }, 0);
-  let columnSequence = rawColumns.reduce((max, rawColumn) => {
-    return typeof rawColumn.id === "string"
-      ? Math.max(max, numericSuffix(rawColumn.id.trim(), "c"))
-      : max;
-  }, 0);
-  const usedPaneIds = new Set<string>();
-  const usedColumnIds = new Set<string>();
-  const columns: ChatSplitColumn[] = [];
-  const sourceColumnIndexes: number[] = [];
-  for (const [columnIndex, rawColumn] of rawColumns.entries()) {
-    if (!Array.isArray(rawColumn.panes)) {
-      continue;
-    }
-    const panes: ChatSplitPane[] = [];
-    const sourcePaneIndexes: number[] = [];
-    for (const [paneIndex, rawPane] of rawColumn.panes.entries()) {
-      if (!isRecord(rawPane) || typeof rawPane.sessionKey !== "string") {
-        continue;
-      }
-      const sessionKey = rawPane.sessionKey.trim();
-      if (!sessionKey) {
-        continue;
-      }
-      panes.push({
-        id: uniqueId(rawPane.id, usedPaneIds, () => `p${++paneSequence}`),
-        sessionKey,
-      });
-      sourcePaneIndexes.push(paneIndex);
-    }
-    if (panes.length === 0) {
-      continue;
-    }
-    const rawPaneWeights = readWeights(rawColumn.paneWeights, rawColumn.panes.length);
-    const paneWeights = normalizedWeights(
-      sourcePaneIndexes.map((index) =>
-        expectDefined(rawPaneWeights[index], "normalized split pane source weight"),
-      ),
-    );
-    columns.push({
-      id: uniqueId(rawColumn.id, usedColumnIds, () => `c${++columnSequence}`),
-      panes,
-      paneWeights,
-    });
-    sourceColumnIndexes.push(columnIndex);
-  }
-  if (columns.length === 0) {
-    return undefined;
-  }
-  const rawColumnWeights = readWeights(value.columnWeights, rawColumns.length);
-  const columnWeights = normalizedWeights(
-    sourceColumnIndexes.map((index) =>
-      expectDefined(rawColumnWeights[index], "normalized split column source weight"),
-    ),
-  );
-  const allPanes = columns.flatMap((column) => column.panes);
-  if (allPanes.length < 2) {
-    return undefined;
-  }
-  const requestedActivePaneId =
-    typeof value.activePaneId === "string" ? value.activePaneId.trim() : "";
-  const activePaneId = allPanes.some((pane) => pane.id === requestedActivePaneId)
-    ? requestedActivePaneId
-    : expectDefined(allPanes[0], "normalized split layout first pane").id;
-  return { columns, columnWeights, activePaneId };
 }
